feat: Add delete endpoint for feedback
Add a new endpoint to delete feedback given the organization ID and feedback ID. The function `delete_feedback()` in `FeedbackRepository` is responsible for deleting the feedback from the database. The endpoint is implemented as a `DELETE` request with a status code of 204. The feedback is deleted by calling the `FeedbackRepository.delete_feedback()` method, passing the organization ID and feedback ID as parameters. After deletion, a check is performed to ensure that the feedback is no longer accessible by making a `GET` request to the same endpoint.